Day 9

+++ AoC2019/Solutions/Day5.cs
@@ -11,7 +11,7 @@ namespace AoC2019.Solutions
     {
         public static string A(string input)
         {
-            var program = input.Split(',').Select(int.Parse).ToList();
+            var program = input.Split(',').Select(long.Parse).ToList();
 
             var computer = new IntCodeComputer(program);
             computer.AddToInput(1);
@@ -22,7 +22,7 @@ namespace AoC2019.Solutions
 
         public static string B(string input)
         {
-            var program = input.Split(',').Select(int.Parse).ToList();
+            var program = input.Split(',').Select(long.Parse).ToList();
             var computer = new IntCodeComputer(program);
             computer.AddToInput(5);
             computer.Run();
